rename file to be shorter
